+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Fix detection of `flutter_driver` for events other than the first.
+- Add locale to `environmentAttributes` in `FlutterSentry.wrap()`.
 
 ## 0.3.1